Add support for clang __nodebug, attempt (and fail) to do something similar for MSVC
I added a crappy env var XHD_DEBUG_STEPOVER that can be used to turn on a useless attempt to inline some stuff in MSVC debug builds. Don't bother. I'll likely delete it later but I'm leaving it in for now jsut to remind me to never try this again. (it's my third time, I need to learn)