Finish cleaning up the windows console hack-a-thon.
Officially gave up on the idea of making Subsystem:Windows apps work in an understandable or consistent manner across various shells and terminals and pipe redirects. This allows me to delete a lot of code that was going nowhere. If we do provide consoles to windowed apps it should be with manual user cli siwtches or such actions - automatically trying to attach to a parent process terminal is just dangerous and too likely to yield unwanted behavior outside of Bash/Mintty.