stat: disambiguate ctime (changed/create) and add CStatModinfo for just modification tracking
Create timestamp exists on windows.
Changed timestamp exists on POSIX.

CStatModinfo cares only about filesize and modification time, which is sufficient to avoid spurious reloads when watching files for changes. Changes to mode of the file nor access or creation/change time are relevant since hot-reloading only cares if the contents of the file have appeared to have changed.