xem_startup: Wait-for-Debugger functionality (WAIT_DEBUG=1 or --wait-debug)
Process enters a sleep wait very early in main(), and resumes execution automatically as soon as the debugger is attached.

This is not to be confused with "Attach to Debugger" functionality which depends on the external too `atd.exe`. By comparison, simply waiting for a debugger requires no external dependencies, but does require the user to manually attach a debugger of their choice to the stopped process.