defer: rename Defer -> defer
Also changes API to use __VA_ARGS__ instead of naked statement, which allows it to accept compound comma-delimited statements in a more natural way.

I had used the uppercase name long ago out of some concern for namespace pollution. But other projects (namely Syrup) are using lowercase defer without issues and so I think it's fine to prefer it. And it more closely aligns with the Golang syntax that we're borrowing from.