Scrub DEBUG, HOUSE, RETAIL from the sources
libimplicitstd is a general base library and shouldn't care about the build type of the final platform.

Later on it might make sense to add a LIBIMPLICIT_DEVELOPER flag, for use only by those actively developing on the library, to enact lots of extra diagnostics or debuggable behaviors. At this time there's no obvious need for it so going to punt that addition down the road until such time it has purpose.